Change font size to `CGFloat`.

+++ Source/model/draw/Font.swift
@@ -3,9 +3,9 @@ import Foundation
 open class Font {
 
 	open let name: String
-	open let size: Int
+	open let size: CGFloat
 
-	public init(name: String = "Serif", size: Int = 12) {
+	public init(name: String = "Serif", size: CGFloat = 12) {
 		self.name = name
 		self.size = size
 	}
